Add ramp-up tests for transport sequence number with and w/o audio.
Also add a perf metric tracking the average network latency.

The audio stream test is disabled for now since audio isn't included in bitrate allocation.

BUG=webrtc:5263
R=mflodman@webrtc.org

Review URL: https://codereview.webrtc.org/1582833002 .

Cr-Commit-Position: refs/heads/master@{#11244}

+++ webrtc/test/fake_network_pipe.cc
@@ -146,7 +146,8 @@ int FakeNetworkPipe::AverageDelay() {
   if (sent_packets_ == 0)
     return 0;
 
-  return total_packet_delay_ / static_cast<int>(sent_packets_);
+  return static_cast<int>(total_packet_delay_ /
+                          static_cast<int64_t>(sent_packets_));
 }
 
 void FakeNetworkPipe::Process() {
